change label data type int32 -> int64

+++ vgg16.py
@@ -52,7 +52,7 @@ def read_labeled_tfrecord(example):
     }
     example = tf.io.parse_single_example(example, LABELED_TFREC_FORMAT)
     image = decode_image(example['image'])
-    label = tf.cast(example['class'], tf.int32)
+    label = tf.cast(example['class'], tf.int64)
     return image, label  # returns a dataset of (image, label) pairs
